chore(zones): removes nav tabs from detail views
