+++ src/app/zones/views/ZoneIndexView.vue
@@ -20,7 +20,7 @@
       </template>
 
       <template
-        v-if="env('KUMA_ZONE_CREATION_FLOW') === 'enabled' && isGlobalMode"
+        v-if="env('KUMA_ZONE_CREATION_FLOW') === 'enabled' && store.getters['config/getMulticlusterStatus']"
         #actions
       >
         <KButton
@@ -31,11 +31,6 @@
           {{ t('zones.index.create') }}
         </KButton>
       </template>
-
-      <NavTabs
-        v-if="store.getters['config/getMulticlusterStatus']"
-        :tabs="tabs"
-      />
 
       <RouterView v-slot="{ Component, route }">
         <component
@@ -49,36 +44,14 @@
 
 <script lang="ts" setup>
 import { KButton } from '@kong/kongponents'
-import { computed } from 'vue'
 
 import AppView from '@/app/application/components/app-view/AppView.vue'
 import RouteTitle from '@/app/application/components/route-view/RouteTitle.vue'
 import RouteView from '@/app/application/components/route-view/RouteView.vue'
-import NavTabs, { NavTab } from '@/app/common/NavTabs.vue'
 import { useStore } from '@/store/store'
 import { useEnv, useI18n } from '@/utilities'
 
 const env = useEnv()
 const { t } = useI18n()
 const store = useStore()
-
-const isGlobalMode = computed(() => store.getters['config/getMulticlusterStatus'])
-
-const tabs: NavTab[] = [
-  {
-    title: t('zones.routes.items.navigation.zone-cp-list-view'),
-    routeName: 'zone-cp-list-view',
-    module: 'zone-cps',
-  },
-  {
-    title: t('zones.routes.items.navigation.zone-ingress-list-view'),
-    routeName: 'zone-ingress-list-view',
-    module: 'zone-ingresses',
-  },
-  {
-    title: t('zones.routes.items.navigation.zone-egress-list-view'),
-    routeName: 'zone-egress-list-view',
-    module: 'zone-egresses',
-  },
-]
 </script>
